Added conditional deleting of previous eulogy

+++ app/Http/Controllers/FilePondController.php
@@ -177,7 +177,7 @@ class FilePondController extends Controller
             $death = Death::findOrFail($id);
 
             // Delete Old Eulogy
-            Storage::disk("public")->delete($death->eulogy);
+            $death->eulogy && Storage::disk("public")->delete($death->eulogy);
 
             $death->eulogy = $eulogy;
